Split test_macros.py
